Refactor entity inheritance and type changes
Changed `Id` and `ArticleId` properties from `int` to `long` in various models. Introduced `MetaEntityBase` abstract class for meta entities. Updated `Article`, `ArticleMeta`, `AuthorMeta`, and `SectionMeta` classes to inherit from new base classes. Modified repository interfaces and implementations to use `long` instead of `int`.

+++ src/Core/Model/ArticleSection.cs
@@ -2,7 +2,7 @@
 
 public sealed class ArticleSection : EntityBase<long>
 {
-	public int ArticleId { get; set; }
+	public long ArticleId { get; set; }
 	public int SectionId { get; set; }
 
 	public Article? Article { get; set; }
